Aktualizace stupnic modelů

+++ Meteo/CloudModelSpectrum.cs
@@ -9,7 +9,7 @@ namespace Meteo
     public class CloudModelSpectrum
     {
         public int id_model { get; set; }
-        public int rank { get; set; }
+        public float rank { get; set; }
         public string color { get; set; }
         public int type { get; set; }
 
@@ -20,7 +20,7 @@ namespace Meteo
         public CloudModelSpectrum(int id_model, string rank, string color, string type)
         {
             this.id_model = id_model;
-            this.rank = Int32.Parse(rank);
+            this.rank = float.Parse(rank);
             this.color = color;
             this.type = Int32.Parse(type);
         }
@@ -29,7 +29,7 @@ namespace Meteo
         public CloudModelSpectrum(string id_model, string rank, string color)
         {
             this.id_model = Int32.Parse(id_model);
-            this.rank = Int32.Parse(rank);
+            this.rank = float.Parse(rank);
             this.color = color;
         }
     }
